// src/types/common.ts
// src/types/common.ts
// Purpose: Common type definitions used across the app

export type Response<T> = {
  result: {
    success: boolean;
    code: string;
    message: string;
  };
  body: T;
  page?: number;
  size?: number;
  total?: number;
  totalPages?: number;
};

export type User = {
  id: number;
  provider: string;
  providerId: string;
  name: string;
  email: string;
  profileImage: string;
};

// Color types
export enum Category {
  ALL = '',
  CHEST = 'CHEST',
  BACK = 'BACK',
  SHOULDER = 'SHOULDER',
  LEG = 'LEG',
  ARM = 'ARM',
  ETC = 'ETC',
}
export enum Color {
  VIOLET = '#855CF8',
  ORANGE = '#F26B2C',
  GREEN = '#2DAF2D',
  BLUE = '#455A64',
  PINK = '#DD408F',
}
export enum BackgroundColor {
  VIOLET = '#F1E6FF', // 연한 보라색
  ORANGE = '#FFE5D1', // 연한 오렌지색
  GREEN = '#F0FFF0', // 연한 초록색
  BLUE = '#E6F5FD', // 연한 파란색
  PINK = '#FFE6F2', // 연한 분홍색
}
export enum Period {
  WEEK = 'WEEK',
  MONTH = 'MONTH',
  QUARTER = 'QUARTER',
  HALF = 'HALF',
  YEAR = 'YEAH',
  ALL = 'ALL',
}
export enum Type {
  WEIGHT = 'WEIGHT',
  COUNT = 'COUNT',
  TIME = 'TIME',
  REST = 'REST', // 휴식 타입 추가
}